Refactor policies into subpackage
