--- ifera/policies.py
@@ -1,884 +1,0 @@
-"""
-Module containing trading policy implementations for making trading decisions.
-
-Trading policies are responsible for determining actions based on market data
-and current trading state. The module provides a composite policy structure
-that delegates decisions to specialized sub-policies depending on the current
-trading context.
-"""
-
-from abc import ABC, abstractmethod
-from typing import List, Tuple
-
-import torch
-from torch import nn
-from einops import rearrange, repeat
-from .data_models import InstrumentData, DataManager
-from .config import ConfigManager
-from .file_manager import FileManager
-
-
-class PositionMaintenancePolicy(nn.Module, ABC):
-    """
-    Abstract base class for position maintenance policies.
-
-    Position maintenance policies are responsible for managing existing positions,
-    including updating stop loss levels and determining when to close positions.
-
-    All implementations must provide reset and forward methods with the specified
-    signatures.
-    """
-
-    @abstractmethod
-    def reset(self, mask: torch.Tensor) -> None:
-        """
-        Reset the policy's state for the specified batch elements.
-
-        Args:
-            mask (torch.Tensor): A boolean tensor where True indicates batches to reset.
-        """
-        pass
-
-    @abstractmethod
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        """
-        Compute actions and stop loss levels for the current state.
-
-        Args:
-            date_idx (torch.Tensor): Batch of date indices.
-            time_idx (torch.Tensor): Batch of time indices.
-            position (torch.Tensor): Current positions (non-zero for existing positions).
-            prev_stop (torch.Tensor): Previous stop loss levels.
-            entry_price (torch.Tensor): Entry prices for the positions.
-
-        Returns:
-            Tuple[torch.Tensor, torch.Tensor]: (action, stop_loss)
-                - action: Actions to take (0 = no action, positive = buy, negative = sell)
-                - stop_loss: Updated stop loss levels.
-        """
-        pass
-
-
-class TradingDonePolicy(nn.Module, ABC):
-    """Abstract base class for trading done policies."""
-
-    @abstractmethod
-    def reset(self, mask: torch.Tensor) -> None:
-        """Reset the policy's state for the specified batch elements."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> torch.Tensor:
-        """Return a boolean tensor indicating completed episodes."""
-        raise NotImplementedError
-
-
-class TradingPolicy(nn.Module):
-    """
-    A composite trading policy that delegates to specialized sub-policies.
-
-    This policy orchestrates multiple sub-policies that handle different aspects
-    of trading decision making:
-    - Opening new positions
-    - Setting initial stop loss levels
-    - Managing existing positions
-    - Determining when episodes should terminate (for backtesting/RL)
-
-    All methods operate on batched inputs and outputs, supporting vectorized
-    operations for efficiency.
-
-    Parameters
-    ----------
-    instrument_data : InstrumentData
-        Data for the financial instrument being traded
-    open_position_policy : object
-        Policy for determining actions when current position is 0
-    initial_stop_loss_policy : object
-        Policy for setting stop loss levels when opening new positions
-    position_maintenance_policy : PositionMaintenancePolicy
-        Policy for managing existing positions (actions and stop loss updates)
-
-    Notes
-    -----
-    All sub-policies must implement a forward method with compatible signatures.
-    """
-
-    def __init__(
-        self,
-        instrument_data: InstrumentData,
-        open_position_policy: torch.nn.Module,
-        initial_stop_loss_policy: torch.nn.Module,
-        position_maintenance_policy: PositionMaintenancePolicy,
-        trading_done_policy: TradingDonePolicy,
-        batch_size: int,
-    ) -> None:
-        super().__init__()
-        self.instrument_data = instrument_data
-        self.open_position_policy = open_position_policy
-        self.initial_stop_loss_policy = initial_stop_loss_policy
-        self.position_maintenance_policy = position_maintenance_policy
-        self.trading_done_policy = trading_done_policy
-        self._batch_size = batch_size
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        """
-        Determine trading actions and stop loss levels based on current state.
-
-        This method delegates to the appropriate sub-policies based on the current
-        position state and already-done episodes.
-
-        Parameters
-        ----------
-        date_idx : torch.Tensor
-            Batch of date indices
-        time_idx : torch.Tensor
-            Batch of time indices
-        position : torch.Tensor
-            Current positions for each batch element (0 = no position)
-        prev_stop : torch.Tensor
-            Previous stop loss levels
-        entry_price : torch.Tensor
-            Entry price for current positions
-
-        Returns
-        -------
-        action : torch.Tensor
-            Actions to take (0 = no action, positive = buy, negative = sell)
-        stop_loss : torch.Tensor
-            Stop loss price levels
-        done : torch.Tensor
-            Boolean tensor indicating finished episodes
-        """
-        # Create masks for different position states
-        no_position_mask = position == 0
-        has_position_mask = position != 0
-
-        # Handle batches with no position (position == 0)
-        action = self.open_position_policy(
-            date_idx,
-            time_idx,
-            no_position_mask,
-        )
-
-        # For batches where we're opening a position, get initial stop loss
-        opening_position_mask = action != 0
-
-        stop_loss = self.initial_stop_loss_policy(
-            date_idx,
-            time_idx,
-            action,
-        )
-        self.position_maintenance_policy.reset(opening_position_mask)
-
-        # Handle batches with existing positions
-        maintenance_actions, maintenance_stops = self.position_maintenance_policy(
-            date_idx, time_idx, position, prev_stop, entry_price
-        )
-        action = torch.where(has_position_mask, maintenance_actions, action)
-        stop_loss = torch.where(has_position_mask, maintenance_stops, stop_loss)
-
-        done = self.trading_done_policy(
-            date_idx, time_idx, position, prev_stop, entry_price
-        )
-        last_bar_mask = (date_idx == self.instrument_data.data.shape[0] - 1) & (
-            time_idx == self.instrument_data.data.shape[1] - 1
-        )
-        done = done | last_bar_mask
-
-        return action, stop_loss, done
-
-
-class AlwaysOpenPolicy(nn.Module):
-    """
-    Policy that always opens a position.
-
-    This policy is used to test the trading policy structure with a simple
-    sub-policy that always opens a position.
-
-    Parameters
-    ----------
-    direction : int
-        Direction of the position (1 = long, -1 = short
-    """
-
-    def __init__(self, direction: int, batch_size: int, device: torch.device) -> None:
-        super().__init__()
-        _ = batch_size
-        self.direction = torch.tensor(direction, dtype=torch.int32, device=device)
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        no_position_mask: torch.Tensor,
-    ) -> torch.Tensor:
-        """
-        Determine actions for opening new positions.
-
-        Parameters
-        ----------
-        date_idx : torch.Tensor
-            Batch of date indices
-        time_idx : torch.Tensor
-            Batch of time indices
-        no_position_mask : torch.Tensor
-            Mask indicating which batch elements have no position (0 = no position)
-
-        Returns
-        -------
-        action : torch.Tensor
-            Actions to take (0 = no action, positive = buy, negative = sell)
-        """
-        _, _ = date_idx, time_idx  # Unused in this policy
-        return no_position_mask * self.direction
-
-
-class OpenOncePolicy(nn.Module):
-    """
-    Policy that opens a position once and holds it.
-
-    This policy is used to test the trading policy structure with a simple
-    sub-policy that opens a position once and holds it.
-
-    Parameters
-    ----------
-    direction : int
-        Direction of the position (1 = long, -1 = short)
-    """
-
-    def __init__(self, direction, batch_size, device) -> None:
-        super().__init__()
-        self.direction = torch.tensor(direction, dtype=torch.int32, device=device)
-        self.opened = torch.zeros((batch_size,), dtype=torch.bool, device=device)
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        no_position_mask: torch.Tensor,
-    ) -> torch.Tensor:
-        """
-        Determine actions for opening new positions.
-
-        Parameters
-        ----------
-        date_idx : torch.Tensor
-            Batch of date indices
-        time_idx : torch.Tensor
-            Batch of time indices
-        no_position_mask : torch.Tensor
-            Mask indicating which batch elements have no position (0 = no position)
-
-        Returns
-        -------
-        action : torch.Tensor
-            Actions to take (0 = no action, positive = buy, negative = sell)
-        """
-        _, _ = date_idx, time_idx  # Unused in this policy
-
-        open_mask = no_position_mask & ~self.opened
-        self.opened = self.opened | open_mask
-
-        return open_mask * self.direction
-
-
-class AlwaysFalseDonePolicy(TradingDonePolicy):
-    """Trading done policy that never signals completion."""
-
-    def __init__(self, batch_size: int, device: torch.device) -> None:
-        super().__init__()
-        self._false = torch.zeros(batch_size, dtype=torch.bool, device=device)
-
-    def reset(self, mask: torch.Tensor) -> None:  # pragma: no cover - no state
-        """Reset policy state (no-op)."""
-        return None
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> torch.Tensor:
-        """Return ``False`` for all batches."""
-        _ = date_idx, time_idx, position, prev_stop, entry_price
-        return self._false
-
-
-class SingleTradeDonePolicy(TradingDonePolicy):
-    """Signals done once a non-zero position returns to zero."""
-
-    def __init__(self, batch_size: int, device: torch.device) -> None:
-        super().__init__()
-        self.had_position = torch.zeros(batch_size, dtype=torch.bool, device=device)
-        self._indices = torch.arange(batch_size, device=device)
-
-    def reset(self, mask: torch.Tensor) -> None:
-        """Reset internal state for the selected batch elements."""
-        self.had_position = torch.where(mask, False, self.had_position)
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> torch.Tensor:
-        """Return True when position was open and is now zero."""
-        _ = date_idx, time_idx, prev_stop, entry_price
-        done = (position == 0) & self.had_position
-        self.had_position = torch.where(position != 0, True, self.had_position)
-        return done
-
-
-class ArtrStopLossPolicy(nn.Module):
-    """
-    Policy for setting stop loss levels based on ATR multiples.
-
-    This policy sets stop loss levels based on the average true range (ATR)
-    of the instrument data. The stop loss is set at a multiple of the ATR
-    from the current price.
-
-    Parameters
-    ----------
-    instrument_data : InstrumentData
-        Data for the financial instrument being traded
-    atr_multiple : float
-        Multiple of the ATR to use for setting stop loss levels
-    alpha : float
-        Smoothing factor for ATR calculation
-    acrossday : bool
-        Whether to calculate ATR across days (True) or within the same day (False)
-    """
-
-    def __init__(
-        self,
-        instrument_data: InstrumentData,
-        atr_multiple: float,
-        alpha: float = 1.0 / 14.0,
-        acrossday: bool = True,
-    ) -> None:
-        super().__init__()
-        self.idata = instrument_data
-        self.atr_multiple = atr_multiple
-        if len(instrument_data.artr) == 0:
-            instrument_data.calculate_artr(alpha=alpha, acrossday=acrossday)
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        action: torch.Tensor,
-        prev_stop: torch.Tensor,
-    ) -> torch.Tensor:
-        """
-        Determine stop loss levels for opening new positions.
-
-        Parameters
-        ----------
-        date_idx : torch.Tensor
-            Batch of date indices
-        time_idx : torch.Tensor
-            Batch of time indices
-        position : torch.Tensor
-            Current positions for each batch element (0 = no
-            position, positive = long, negative = short)
-        action : torch.Tensor
-            Actions to take (0 = no action, positive = buy, negative = sell)
-        prev_stop : torch.Tensor
-            Previous stop loss levels
-
-        Returns
-        -------
-        stop_loss : torch.Tensor
-            Stop loss price levels
-        """
-        direction = (position + action).sign()
-
-        # Replace NaN values in prev_stop with infinity (negative for long positions,
-        # positive for short positions) to ensure proper stop loss initialization
-        prev_stop = torch.where(
-            torch.isnan(prev_stop) & (direction != 0),
-            torch.inf * direction * -1,
-            prev_stop,
-        )
-
-        # Relative ATR multiplier for stop loss levels
-        artr = self.idata.artr[date_idx, time_idx] * self.atr_multiple + 1.0
-
-        # For new positions use the close, for trailing stops use the high/low
-        reference_channel = torch.where(
-            position == 0, 3, torch.where(direction > 0, 1, 2)
-        )
-        reference_price = self.idata.data[date_idx, time_idx, reference_channel]
-
-        stop_price = torch.where(
-            direction > 0,
-            torch.maximum(prev_stop, reference_price / artr),
-            torch.minimum(prev_stop, reference_price * artr),
-        )
-        stop_price = torch.where(
-            torch.isnan(stop_price) | (direction == 0), prev_stop, stop_price
-        )
-
-        return stop_price
-
-
-class InitialArtrStopLossPolicy(nn.Module):
-    """
-    Policy for setting initial stop loss levels based on ATR multiples.
-
-    This policy sets stop loss levels based on the average true range (ATR)
-    of the instrument data. The stop loss is set at a multiple of the ATR
-    from the current price.
-
-    Parameters
-    ----------
-    instrument_data : InstrumentData
-        Data for the financial instrument being traded
-    atr_multiple : float
-        Multiple of the ATR to use for setting stop loss levels
-    """
-
-    def __init__(
-        self, instrument_data: InstrumentData, atr_multiple: float, batch_size: int
-    ) -> None:
-        super().__init__()
-        self.artr_policy = ArtrStopLossPolicy(instrument_data, atr_multiple)
-        dtype = instrument_data.data.dtype
-        device = instrument_data.device
-        self._zero = torch.zeros(batch_size, dtype=torch.int32, device=device)
-        self._nan = torch.full((batch_size,), float("nan"), dtype=dtype, device=device)
-
-    def forward(
-        self, date_idx: torch.Tensor, time_idx: torch.Tensor, action: torch.Tensor
-    ) -> torch.Tensor:
-        """
-        Determine stop loss levels for opening new positions.
-
-        Parameters
-        ----------
-        date_idx : torch.Tensor
-            Batch of date indices
-        time_idx : torch.Tensor
-            Batch of time indices
-        action : torch.Tensor
-            Actions to take (0 = no action, positive = buy, negative = sell)
-
-        Returns
-        -------
-        stop_loss : torch.Tensor
-            Stop loss price levels
-        """
-        return self.artr_policy(date_idx, time_idx, self._zero, action, self._nan)
-
-
-class ScaledArtrMaintenancePolicy(PositionMaintenancePolicy):
-    """
-    A position maintenance policy that adjusts stop loss levels using scaled ARTR.
-
-    This policy uses multiple stages with increasing time intervals to calculate
-    stop loss levels based on ARTR. It progresses through stages when certain
-    conditions are met, such as achieving breakeven or sufficient improvement
-    in the stop loss level.
-
-    Attributes:
-        instrument_data (InstrumentData): Data for the financial instrument.
-        stages (List[str]): List of time intervals (e.g., ["1m", "5m", "1h"]).
-        atr_multiple (float): Multiplier for ARTR in stop loss calculations.
-        wait_for_breakeven (bool): Whether to wait for breakeven before advancing stages.
-        minimum_improvement (float): Minimum improvement ratio for stage progression.
-        stage (torch.Tensor): Current stage for each batch element.
-        base_price (torch.Tensor): Base price for improvement calculations.
-    """
-
-    def __init__(
-        self,
-        instrument_data: InstrumentData,
-        stages: List[str],
-        atr_multiple: float,
-        wait_for_breakeven: bool,
-        minimum_improvement: float,
-        batch_size: int,
-    ) -> None:
-        super().__init__()
-        self.instrument_data = instrument_data
-        self.atr_multiple = atr_multiple
-        self.wait_for_breakeven = wait_for_breakeven
-        self.minimum_improvement = minimum_improvement
-
-        cm = ConfigManager()
-        dm = DataManager()
-        fm = FileManager()
-
-        if stages[0] != instrument_data.instrument.interval:
-            raise ValueError("The first stage must match the instrument's interval.")
-
-        self.derived_configs = [
-            cm.get_base_instrument_config(
-                symbol=instrument_data.instrument.symbol,
-                interval=stage,
-                contract_code=instrument_data.instrument.contract_code,
-            )
-            for stage in stages
-        ]
-
-        with fm.persistentContext():
-            self.derived_data = [
-                dm.get_instrument_data(
-                    config,
-                    dtype=instrument_data.dtype,
-                    device=instrument_data.device,
-                    backadjust=instrument_data.backadjust,
-                )
-                for config in self.derived_configs
-            ]
-
-        date_idx_base = torch.arange(
-            self.derived_data[0].data.size(0), device=instrument_data.device
-        )
-        time_idx_base = torch.arange(
-            self.derived_data[0].data.size(1), device=instrument_data.device
-        )
-        date_idx = repeat(date_idx_base, "d -> d t", t=time_idx_base.size(0))
-        time_idx = repeat(time_idx_base, "t -> d t", d=date_idx_base.size(0))
-
-        self.converted_indices = [
-            self.derived_data[s].convert_indices(
-                self.derived_data[0], date_idx, time_idx
-            )
-            for s in range(len(self.derived_data))
-        ]
-
-        self.artr_policies = [
-            ArtrStopLossPolicy(data, self.atr_multiple) for data in self.derived_data
-        ]
-        self.stage_count = len(stages)
-
-        # Initialize state tensors
-        device = instrument_data.device
-        dtype = instrument_data.dtype
-        self.stage = torch.zeros(batch_size, dtype=torch.long, device=device)
-        self.base_price = torch.full(
-            (batch_size,), float("nan"), dtype=dtype, device=device
-        )
-        self._action = torch.zeros(batch_size, dtype=torch.int32, device=device)
-        self._stop = torch.empty(batch_size, dtype=dtype, device=device)
-        self._zero = torch.zeros(batch_size, dtype=torch.int32, device=device)
-        self._nan = torch.full((batch_size,), float("nan"), dtype=dtype, device=device)
-
-    def reset(self, mask: torch.Tensor) -> None:
-        """
-        Reset the policy's state tensors for a given batch size.
-
-        Args:
-            mask (torch.Tensor): A boolean tensor where True indicates batches to reset.
-        """
-        self.stage = torch.where(mask, self._zero, self.stage)
-        self.base_price = torch.where(mask, self._nan, self.base_price)
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Compute actions using scaled ARTR policy."""
-
-        # Initialize outputs
-        action = self._action
-        stop_loss = prev_stop.clone()
-
-        # Extract current state for the subset
-        has_position_mask = position != 0
-        current_stage = self.stage
-        current_base_price = self.base_price.clone()
-
-        # Set base_price
-        nan_base_mask = torch.isnan(self.base_price) & has_position_mask
-        if self.wait_for_breakeven:
-            current_base_price = torch.where(
-                nan_base_mask, entry_price, current_base_price
-            )
-        else:
-            finite_prev_stop = torch.isfinite(prev_stop)
-            set_mask = nan_base_mask & finite_prev_stop
-            current_base_price = torch.where(set_mask, prev_stop, current_base_price)
-
-        # Handle stage 0
-        stage0_mask = current_stage == 0
-        if self.wait_for_breakeven:
-            potential_stop = self.artr_policies[0](
-                date_idx,
-                time_idx,
-                position * stage0_mask,
-                self._zero,
-                stop_loss,
-            )
-            improve_mask_subset = stage0_mask & (
-                (position > 0) & (potential_stop > entry_price)
-                | (position < 0) & (potential_stop < entry_price)
-            )
-            stop_loss = torch.where(improve_mask_subset, potential_stop, stop_loss)
-            current_stage = torch.where(improve_mask_subset, 1, current_stage)
-        else:
-            current_stage = torch.where(
-                stage0_mask & has_position_mask, 1, current_stage
-            )
-
-        # Handle stages > 0
-        for s in range(1, self.stage_count):
-            stage_mask = current_stage == s
-            conv_date_idx = self.converted_indices[s][0][date_idx, time_idx]
-            conv_time_idx = self.converted_indices[s][1][date_idx, time_idx]
-            potential_stop = self.artr_policies[s](
-                conv_date_idx,
-                conv_time_idx,
-                position * stage_mask,
-                self._zero,
-                stop_loss,
-            )
-            improvement = torch.where(
-                position > 0,
-                potential_stop - stop_loss,
-                stop_loss - potential_stop,
-            )
-            min_improvement = self.minimum_improvement * torch.abs(
-                current_base_price - stop_loss
-            )
-            improve_mask_subset = (
-                stage_mask & (improvement > min_improvement) & (conv_date_idx >= 0)
-            )
-            stop_loss = torch.where(improve_mask_subset, potential_stop, stop_loss)
-            current_stage = torch.where(
-                improve_mask_subset & (s < self.stage_count - 1),
-                s + 1,
-                current_stage,
-            )
-
-        # Update the full state with the modified subset
-        self.stage = current_stage
-        self.base_price = current_base_price
-
-        return action, stop_loss
-
-
-class PercentGainMaintenancePolicy(PositionMaintenancePolicy):
-    """
-    A position maintenance policy that adjusts stop-loss levels to maintain
-    a fixed percentage of unrealized gains.
-
-    This policy operates in two stages:
-    - **Stage 1**: Waits until the price moves N ATR into profit using
-      :class:`ArtrStopLossPolicy`. Transitions to Stage 2 when the
-      ATR-based stop-loss exceeds the entry price (higher for long,
-      lower for short).
-    - **Stage 2**: Sets the stop-loss to retain a fixed percentage of
-      unrealized gains relative to an anchor price, acting as a
-      trailing stop.
-
-    Attributes:
-        instrument_data (InstrumentData): Financial instrument data.
-        stage1_atr_multiple (float): ATR multiple for Stage 1 stop-loss calculation.
-        trailing_stop (bool): Whether to adjust stop-loss as a trailing stop in Stage 1.
-        skip_stage1 (bool): If True, skips Stage 1 and starts at Stage 2.
-        keep_percent (float): Percentage of unrealized gains to retain (0 < keep_percent < 1).
-        anchor_type (str): Anchor price type
-            ("entry", "initial_stop", "artificial", "last_stage1_stop",
-            "artificial_stage2").
-        artr_policy (ArtrStopLossPolicy): Policy for ATR-based stop-loss in Stage 1.
-        stage (torch.Tensor): Current stage per batch element (0 = Stage 1, 1 = Stage 2).
-        anchor (torch.Tensor): Anchor price per batch element for Stage 2 calculations.
-    """
-
-    def __init__(
-        self,
-        instrument_data: InstrumentData,
-        stage1_atr_multiple: float,
-        trailing_stop: bool,
-        skip_stage1: bool,
-        keep_percent: float,
-        anchor_type: str,
-        batch_size: int,
-    ) -> None:
-        super().__init__()
-        # Validate anchor_type and conditions
-        if anchor_type not in [
-            "entry",
-            "initial_stop",
-            "artificial",
-            "last_stage1_stop",
-            "artificial_stage2",
-        ]:
-            raise ValueError("Invalid anchor_type")
-        if anchor_type in ["last_stage1_stop", "artificial_stage2"] and (
-            skip_stage1 or not trailing_stop
-        ):
-            raise ValueError(
-                "anchor_type 'last_stage1_stop' and 'artificial_stage2' require "
-                "skip_stage1=False and trailing_stop=True"
-            )
-        if anchor_type in ["artificial", "artificial_stage2"] and not (
-            0 < keep_percent < 1
-        ):
-            raise ValueError(
-                "keep_percent must be between 0 and 1 for 'artificial' and "
-                "'artificial_stage2' anchor_types"
-            )
-
-        self.instrument_data = instrument_data
-        self.trailing_stop = trailing_stop
-        self.skip_stage1 = skip_stage1
-        self.keep_percent = keep_percent
-        self.anchor_type = anchor_type
-        self.artr_policy = ArtrStopLossPolicy(instrument_data, stage1_atr_multiple)
-
-        device = instrument_data.device
-        dtype = instrument_data.dtype
-        initial_stage = 1 if self.skip_stage1 else 0
-        self.stage = torch.full(
-            (batch_size,), initial_stage, dtype=torch.long, device=device
-        )
-        self.anchor = torch.full(
-            (batch_size,), float("nan"), dtype=dtype, device=device
-        )
-        self._action = torch.zeros(batch_size, dtype=torch.int32, device=device)
-        self._stop = torch.empty(batch_size, dtype=dtype, device=device)
-        self._initial_stage = torch.full(
-            (batch_size,), initial_stage, dtype=torch.long, device=device
-        )
-        self._nan = torch.full((batch_size,), float("nan"), dtype=dtype, device=device)
-
-    def reset(self, mask: torch.Tensor) -> None:
-        """
-        Reset the policy's state for specified batch elements.
-
-        Args:
-            mask (torch.Tensor): Boolean tensor where True indicates batches to reset.
-        """
-        self.stage = torch.where(mask, self._initial_stage, self.stage)
-        self.anchor = torch.where(mask, self._nan, self.anchor)
-
-    def forward(
-        self,
-        date_idx: torch.Tensor,
-        time_idx: torch.Tensor,
-        position: torch.Tensor,
-        prev_stop: torch.Tensor,
-        entry_price: torch.Tensor,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Compute actions using percent gain policy."""
-
-        # Initialize outputs without allocation
-        action = self._action
-        stop_loss = prev_stop.clone()
-
-        # Get current state for the batch
-        has_position_mask = position != 0
-        current_stage = self.stage.clone()
-
-        # Set anchor for certain anchor_types on first call
-        first_call_mask = torch.isnan(self.anchor) & has_position_mask
-        if self.anchor_type == "entry":
-            self.anchor = torch.where(first_call_mask, entry_price, self.anchor)
-        elif self.anchor_type == "initial_stop":
-            self.anchor = torch.where(first_call_mask, prev_stop, self.anchor)
-        elif self.anchor_type == "artificial":
-            reference_channel = torch.where(position > 0, 1, 2)  # 1 for high, 2 for low
-            reference_price = self.instrument_data.data[
-                date_idx, time_idx, reference_channel
-            ]
-            self.anchor = torch.where(
-                first_call_mask,
-                (prev_stop - self.keep_percent * reference_price)
-                / (1 - self.keep_percent),
-                self.anchor,
-            )
-
-        # Stage 1: ATR-based waiting period
-        stage1_mask = (current_stage == 0) & has_position_mask
-        atr_stop = self.artr_policy(
-            date_idx,
-            time_idx,
-            position * stage1_mask,
-            action,
-            prev_stop,
-        )
-        move_to_stage2 = stage1_mask & (
-            (position > 0) & (atr_stop > entry_price)
-            | (position < 0) & (atr_stop < entry_price)
-        )
-
-        if self.anchor_type == "last_stage1_stop":
-            self.anchor = torch.where(move_to_stage2, atr_stop, self.anchor)
-        elif self.anchor_type == "artificial_stage2":
-            reference_channel = torch.where(position > 0, 1, 2)
-            reference_price = self.instrument_data.data[
-                date_idx,
-                time_idx,
-                reference_channel,
-            ]
-            self.anchor = torch.where(
-                move_to_stage2,
-                (atr_stop - self.keep_percent * reference_price)
-                / (1 - self.keep_percent),
-                self.anchor,
-            )
-
-        self.stage = torch.where(move_to_stage2, 1, current_stage)
-
-        if self.trailing_stop:
-            stop_loss = torch.where(stage1_mask, atr_stop, stop_loss)
-
-        # Stage 2: Percent-based stop-loss
-        stage2_mask = current_stage == 1
-        anchor = self.anchor
-        # Use high for long, low for short as reference price
-        reference_channel = torch.where(position > 0, 1, 2)
-        reference_price = self.instrument_data.data[
-            date_idx, time_idx, reference_channel
-        ]
-        # Calculate candidate stop-loss
-        candidate_stop_loss = torch.where(
-            position > 0,
-            torch.maximum(
-                prev_stop, anchor + self.keep_percent * (reference_price - anchor)
-            ),
-            torch.minimum(
-                prev_stop, anchor - self.keep_percent * (anchor - reference_price)
-            ),
-        )
-        # Ensure stop-loss only tightens (trailing stop behavior)
-        stop_loss = torch.where(
-            stage2_mask,
-            candidate_stop_loss,
-            stop_loss,
-        )
-
-        return action, stop_loss
