+++ ifera/policies/__init__.py
@@ -1,0 +1,36 @@
+"""Convenience exports for policy classes."""
+
+from .trading_policy import BaseTradingPolicy, TradingPolicy
+from .open_position_policy import OpenPositionPolicy, AlwaysOpenPolicy, OpenOncePolicy
+from .stop_loss_policy import (
+    StopLossPolicy,
+    ArtrStopLossPolicy,
+    InitialArtrStopLossPolicy,
+)
+from .position_maintenance_policy import (
+    PositionMaintenancePolicy,
+    ScaledArtrMaintenancePolicy,
+    PercentGainMaintenancePolicy,
+)
+from .trading_done_policy import (
+    TradingDonePolicy,
+    AlwaysFalseDonePolicy,
+    SingleTradeDonePolicy,
+)
+
+__all__ = [
+    "BaseTradingPolicy",
+    "TradingPolicy",
+    "OpenPositionPolicy",
+    "AlwaysOpenPolicy",
+    "OpenOncePolicy",
+    "StopLossPolicy",
+    "ArtrStopLossPolicy",
+    "InitialArtrStopLossPolicy",
+    "PositionMaintenancePolicy",
+    "ScaledArtrMaintenancePolicy",
+    "PercentGainMaintenancePolicy",
+    "TradingDonePolicy",
+    "AlwaysFalseDonePolicy",
+    "SingleTradeDonePolicy",
+]
